# Conflicts:
#	src/BackgroundTasks/ExceptionHandler.php
#	src/BackgroundTasks/Exceptions/NoObserverForUserInteractionException.php
#	src/BackgroundTasks/Implementation/Tasks/SlowPlusJob.php
#	src/BackgroundTasks/Implementation/Values/ScalarValues/ScalarValueFactory.php
#	src/BackgroundTasks/Types/Exceptions/InvalidArgumentException.php

+++ Modules/ContentPage/classes/class.ilContentPageImporter.php
@@ -5,17 +5,11 @@ use ILIAS\ContentPage\PageMetrics\PageMetricsService;
 use ILIAS\ContentPage\PageMetrics\PageMetricsRepositoryImp;
 use ILIAS\ContentPage\PageMetrics\Command\StorePageMetricsCommand;
 
-/**
- * Class ilContentPageImporter
- */
 class ilContentPageImporter extends ilXmlImporter implements ilContentPageObjectConstants
 {
     protected ilContentPageDataSet $ds;
     private PageMetricsService $pageMetricsService;
 
-    /**
-     * @inheritdoc
-     */
     public function init() : void
     {
         global $DIC;
@@ -30,24 +24,18 @@ class ilContentPageImporter extends ilXmlImporter implements ilContentPageObject
         );
     }
 
-    /**
-     * @inheritdoc
-     */
     public function importXmlRepresentation(string $a_entity, string $a_id, string $a_xml, ilImportMapping $a_mapping) : void
     {
         $parser = new ilDataSetImportParser($a_entity, $this->getSchemaVersion(), $a_xml, $this->ds, $a_mapping);
     }
 
-    /**
-     * @inheritdoc
-     */
     public function finalProcessing(ilImportMapping $a_mapping) : void
     {
         parent::finalProcessing($a_mapping);
 
         $copaMap = $a_mapping->getMappingsOfEntity('Services/COPage', 'pg');
         foreach ($copaMap as $oldCopaId => $newCopaId) {
-            $newCopaId = substr($newCopaId, strlen(self::OBJ_TYPE) + 1);
+            $newCopaId = (int) substr($newCopaId, strlen(self::OBJ_TYPE) + 1);
 
             ilContentPagePage::_writeParentId(self::OBJ_TYPE, $newCopaId, $newCopaId);
 
@@ -55,7 +43,7 @@ class ilContentPageImporter extends ilXmlImporter implements ilContentPageObject
             foreach ($translations as $language) {
                 $this->pageMetricsService->store(
                     new StorePageMetricsCommand(
-                        (int) $newCopaId,
+                        $newCopaId,
                         $language
                     )
                 );
@@ -66,7 +54,7 @@ class ilContentPageImporter extends ilXmlImporter implements ilContentPageObject
         foreach ($styleMapping as $newCopaId => $oldStyleId) {
             $newStyleId = (int) $a_mapping->getMapping('Services/Style', 'sty', $oldStyleId);
             if ($newCopaId > 0 && $newStyleId > 0) {
-                $copa = ilObjectFactory::getInstanceByObjId($newCopaId, false);
+                $copa = ilObjectFactory::getInstanceByObjId((int) $newCopaId, false);
                 if (!$copa || !($copa instanceof ilObjContentPage)) {
                     continue;
                 }
